requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
python-multipart>=0.0.12
openai>=1.59
pydantic>=2.10
python-docx>=1.1
PyMuPDF>=1.24
reportlab>=4.0
qhchina
